sanitize package names
