+++ src/main/java/com/calculatorProject/calculatorproject/Calculator.java
@@ -1,4 +1,4 @@
-package com.calculatorProject.calculatorProject;
+package com.calculatorProject.calculatorproject;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,8 +8,6 @@ import lombok.Setter;
 import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -39,8 +37,6 @@ public class Calculator {
         try {
             switch (operator) {
                 case "add":
-//                    mapError.put("code", "200");
-//                    mapError.put("message", "OK");
                     mapData.put("result:",
                             new BigInteger(new BigInteger(numberOne).add(new BigInteger(numberTwo)).toString()).toString());
                     response.setData(mapData);
@@ -48,8 +44,6 @@ public class Calculator {
                     response.setError(null);
                     return response;
                 case "subtract":
-//                    mapError.put("code", "200");
-//                    mapError.put("message", "OK");
                     mapData.put("result:",
                             new BigInteger(new BigInteger(numberOne).subtract(new BigInteger(numberTwo)).toString()).toString());
                     response.setData(mapData);
@@ -57,18 +51,13 @@ public class Calculator {
                     response.setError(null);
                     return response;
                 case "multiply":
-//                    mapError.put("code", "200");
-//                    mapError.put("message", "OK");
                     mapData.put("result:",
                             new BigInteger(new BigInteger(numberOne).multiply(new BigInteger(numberTwo)).toString()).toString());
                     response.setData(mapData);
                     response.setStatus(true);
                     response.setError(null);
                     return response;
-                case "divide": // TODO implement divide method
-//                    mapError.put("code", "200");
-//                    mapError.put("message", "OK");
-
+                case "divide":
                     if (numberTwo.equals("0")) {
                         mapData.put("result:", "Can't divide by zero");
 
@@ -88,8 +77,6 @@ public class Calculator {
                     }
 
                 default:
-//                    mapError.put("code", "200");
-//                    mapError.put("message", "OK");
                     mapData.put("result:", "Invalid Input Data");
                     response.setData(mapData);
                     response.setStatus(false);
@@ -105,31 +92,6 @@ public class Calculator {
             response.setError(null);
             return response;
         }
-    }
-
-    public String calculate(String aritmeticExpression) {
-        String operation;
-        String[] numbers;
-
-
-        String regexNumber = "(\\d+)";
-        String regexNumberOperatorNumber = "(\\d+)\\+(\\d+)";
-
-
-        Matcher matcher = Pattern.compile(regexNumberOperatorNumber).matcher(aritmeticExpression);
-
-        while (matcher.find()) {
-            operation = matcher.group();
-
-            matcher = Pattern.compile(regexNumber).matcher(operation);
-
-//            while (matcher.find()) {
-//                System.out.println(matcher.group(1));
-//                System.out.println(matcher.group(2));
-//            }
-
-        }
-        return "";
     }
 
 }
