+++ src/main/java/com/calculatorProject/calculatorproject/ServletInitializer.java
@@ -1,4 +1,4 @@
-package com.calculatorProject.calculatorProject;
+package com.calculatorProject.calculatorproject;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
