+++ src/test/java/com/calculatorProject/calculatorproject/CalculatorProjectApplicationTests.java
@@ -1,4 +1,4 @@
-package com.calculatorProject.calculatorProject;
+package com.calculatorProject.calculatorproject;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -6,15 +6,19 @@ import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import java.util.HashMap;
-import java.util.Map;
-
 @RunWith(SpringRunner.class)
 @SpringBootTest
 public class CalculatorProjectApplicationTests {
 
     @Test
     public void contextLoads() {
+    }
+
+
+    @Test
+    public void addTwoNumbersTest() {
+        Calculator calculator = new Calculator();
+        Assert.assertEquals(calculator.operate("23333333333333", "3", "add"), "23333333333336");
     }
 
 //    @Test
